+++ src/core/v2/mod.rs
@@ -1,4 +1,5 @@
 //! Core v2 namespace handle module.
 
+pub mod calc;
 pub mod handle;
 pub use handle::CoreV2;